@@ -1,13 +1,13 @@
 group 'com.projectcitybuild'
-version '2.5.5'
+version '3.0.0'
 
 wrapper {
-    gradleVersion = '5.5.1'
+    gradleVersion = '7.0'
     distributionUrl = distributionUrl.replace("bin", "all")
 }
 
 buildscript {
-    ext.kotlin_version = '1.3.72'
+    ext.kotlin_version = '1.5.30'
 
     repositories {
         mavenCentral()
@@ -32,33 +32,37 @@ repositories {
 }
 
 dependencies {
-    compile "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
     compileOnly 'org.spigotmc:spigot-api:1.12.2-R0.1-SNAPSHOT'
     compileOnly 'net.luckperms:api:5.3'
-    implementation 'com.squareup.retrofit2:retrofit:2.5.0'
-    implementation'com.squareup.retrofit2:converter-gson:2.5.0'
-    implementation 'com.squareup.okhttp3:logging-interceptor:3.5.0'
+    implementation 'com.squareup.retrofit2:retrofit:2.9.0'
+    implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
+    implementation 'com.squareup.okhttp3:logging-interceptor:4.9.1'
+    implementation("com.github.shynixn.mccoroutine:mccoroutine-bukkit-api:1.5.0")
+    implementation("com.github.shynixn.mccoroutine:mccoroutine-bukkit-core:1.5.0")
+    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.5.1")
+    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-jdk8:1.5.1")
     testImplementation 'junit:junit:4.13'
     testImplementation 'org.mockito:mockito-core:2.+'
     testImplementation 'com.squareup.okhttp3:mockwebserver:3.5.0'
 }
 
 compileKotlin {
-    kotlinOptions.jvmTarget = "1.8"
+    kotlinOptions.jvmTarget = "16"
 }
 
 compileTestKotlin {
-    kotlinOptions.jvmTarget = "1.8"
+    kotlinOptions.jvmTarget = "16"
 }
 
-test {
-    useJUnit()
-
-    filter {
-        includeTestsMatching "com.projectcitybuild.tests.*"
-    }
-
-    maxHeapSize = '1G'
-}
+//test {
+//    useJUnit()
+//
+//    filter {
+//        includeTestsMatching "com.projectcitybuild.tests.*"
+//    }
+//
+//    maxHeapSize = '1G'
+//}
 
 build.dependsOn shadowJar
