@@ -3,7 +3,7 @@ package com.projectcitybuild.core.network
 import com.projectcitybuild.core.network.mojang.client.MojangClient
 import com.projectcitybuild.core.network.pcb.client.PCBClient
 
-class NetworkClients(
+class APIRequestFactory(
         val pcb: PCBClient,
         val mojang: MojangClient
 )
