@@ -1,7 +1,6 @@
 package com.projectcitybuild.core.network.mojang.requests
 
 import com.projectcitybuild.core.entities.models.MojangPlayer
-import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Path
 import retrofit2.http.Query
@@ -9,9 +8,9 @@ import retrofit2.http.Query
 interface MojangApiInterface {
 
     @GET("users/profiles/minecraft/{username}")
-    fun getMojangPlayer(
+    suspend fun getMojangPlayer(
             @Path("username") playerName: String,
             @Query("at") timestamp: Long? = null
-    ) : Call<MojangPlayer>
+    ) : MojangPlayer?
 
 }
