@@ -3,7 +3,6 @@ package com.projectcitybuild.core.network.pcb.requests
 import com.projectcitybuild.core.entities.models.ApiResponse
 import com.projectcitybuild.core.entities.models.AuthPlayerGroups
 import com.projectcitybuild.core.entities.models.AuthURL
-import retrofit2.Call
 import retrofit2.http.*
 
 interface AuthApiInterface {
@@ -14,16 +13,16 @@ interface AuthApiInterface {
      */
     @FormUrlEncoded
     @POST("auth/minecraft")
-    fun getVerificationUrl(
+    suspend fun getVerificationUrl(
             @Field("minecraft_uuid") uuid: String
-    ) : Call<ApiResponse<AuthURL>>
+    ) : ApiResponse<AuthURL>
 
     /**
      * Fetches the groups that the given UUID belongs to
      */
     @GET("auth/minecraft/{uuid}")
-    fun getUserGroups(
+    suspend fun getUserGroups(
             @Path(value = "uuid") uuid: String
-    ) : Call<ApiResponse<AuthPlayerGroups>>
+    ) : ApiResponse<AuthPlayerGroups>
 
 }
