@@ -1,7 +1,6 @@
 package com.projectcitybuild.core.network.pcb.requests
 
 import com.projectcitybuild.core.entities.models.*
-import retrofit2.Call
 import retrofit2.http.Field
 import retrofit2.http.FormUrlEncoded
 import retrofit2.http.POST
@@ -10,7 +9,7 @@ interface BanApiInterface {
 
     @FormUrlEncoded
     @POST("bans/store/ban")
-    fun storeBan(
+    suspend fun storeBan(
             @Field("player_id") playerId: String,
             @Field("player_id_type") playerIdType: String,
             @Field("player_alias") playerAlias: String,
@@ -19,22 +18,22 @@ interface BanApiInterface {
             @Field("reason") reason: String?,
             @Field("expires_at") expiresAt: Long?,
             @Field("is_global_ban") isGlobalBan: Int
-    ) : Call<ApiResponse<GameBan>>
+    ) : ApiResponse<GameBan>
 
     @FormUrlEncoded
     @POST("bans/store/unban")
-    fun storeUnban(
+    suspend fun storeUnban(
             @Field("player_id") playerId: String,
             @Field("player_id_type") playerIdType: String,
             @Field("staff_id") staffId: String?,
             @Field("staff_id_type") staffIdType: String
-    ) : Call<ApiResponse<GameUnban>>
+    ) : ApiResponse<GameUnban>
 
     @FormUrlEncoded
     @POST("bans/status")
-    fun requestStatus(
+    suspend fun requestStatus(
             @Field("player_id") playerId: String,
             @Field("player_id_type") playerType: String
-    ) : Call<ApiResponse<GameBan>>
+    ) : ApiResponse<GameBan>
 
 }
