@@ -1,29 +1,35 @@
 package com.projectcitybuild.modules.players
 
-import com.projectcitybuild.core.network.NetworkClients
-import com.projectcitybuild.core.contracts.EnvironmentProvider
+import com.projectcitybuild.core.entities.Failure
+import com.projectcitybuild.core.entities.models.ApiError
+import com.projectcitybuild.core.entities.Result
+import com.projectcitybuild.core.entities.Success
+import com.projectcitybuild.core.network.APIRequestFactory
 import com.projectcitybuild.core.entities.models.MojangPlayer
+import com.projectcitybuild.core.network.APIClient
+import com.projectcitybuild.core.network.APIResult
 
 class GetMojangPlayerAction(
-        private val environment: EnvironmentProvider,
-        private val networkClients: NetworkClients
+        private val apiRequestFactory: APIRequestFactory,
+        private val apiClient: APIClient
 ) {
-    sealed class Result {
-        class SUCCESS(val player: MojangPlayer) : Result()
-        class FAILED(val reason: Failure) : Result()
+    sealed class FailReason {
+        class HTTPError(error: ApiError?): FailReason()
+        object NetworkError: FailReason()
+        object PlayerNotFound: FailReason()
     }
 
-    enum class Failure {
-        DESERIALIZE_FAILED,
-    }
+    suspend fun execute(playerName: String, at: Long? = null): Result<MojangPlayer, FailReason> {
+        val mojangApi = apiRequestFactory.mojang.mojangApi
+        val response = apiClient.execute { mojangApi.getMojangPlayer(playerName, timestamp = at) }
 
-    fun execute(playerName: String, at: Long? = null) : Result {
-        val mojangApi = networkClients.mojang.mojangApi
-
-        val request = mojangApi.getMojangPlayer(playerName, timestamp = at)
-        val response = request.execute()
-        val player = response.body() ?: return Result.FAILED(reason = Failure.DESERIALIZE_FAILED)
-
-        return Result.SUCCESS(player)
+        return when (response) {
+            is APIResult.HTTPError -> Failure(FailReason.HTTPError(response.error))
+            is APIResult.NetworkError -> Failure(FailReason.NetworkError)
+            is APIResult.Success -> {
+                if (response.value == null) Failure(FailReason.PlayerNotFound)
+                else Success(response.value)
+            }
+        }
     }
 }
