@@ -1,43 +1,48 @@
 package com.projectcitybuild.platforms.spigot
 
-import com.projectcitybuild.core.contracts.CommandDelegatable
-import com.projectcitybuild.core.contracts.CommandResult
-import com.projectcitybuild.core.contracts.Commandable
-import com.projectcitybuild.core.contracts.EnvironmentProvider
+import com.github.shynixn.mccoroutine.SuspendingCommandExecutor
+import com.github.shynixn.mccoroutine.setSuspendingExecutor
+import com.projectcitybuild.core.contracts.*
 import com.projectcitybuild.core.entities.CommandInput
-import com.projectcitybuild.core.entities.LogLevel
+import com.projectcitybuild.core.entities.CommandResult
+import org.bukkit.command.Command
+import org.bukkit.command.CommandSender
 import org.bukkit.entity.Player
 import org.bukkit.plugin.java.JavaPlugin
-import java.lang.ref.WeakReference
 
 class SpigotCommandDelegate constructor(
-        private val plugin: WeakReference<JavaPlugin>,
-        private val environment: EnvironmentProvider
-): CommandDelegatable {
+        private val plugin: JavaPlugin,
+        private val logger: LoggerProvider
+) {
+    fun register(command: Commandable) {
+        val aliases = command.aliases.plus(command.label)
 
-    override fun register(command: Commandable) {
-        command.aliases.plus(command.label).forEach { alias ->
-            val plugin = plugin.get() ?: return
-
-            plugin.getCommand(alias)?.setExecutor { sender, _, _, args ->
-                try {
-                    val input = CommandInput(
-                            sender = sender,
-                            args = args ?: arrayOf(),
-                            isConsole = sender !is Player
-                    )
-                    when (command.execute(input)) {
-                        CommandResult.EXECUTED -> true
-                        CommandResult.INVALID_INPUT -> false
+        aliases.forEach { alias ->
+            class BridgedCommand(private val wrappedCommand: Commandable): SuspendingCommandExecutor {
+                override suspend fun onCommand(sender: CommandSender, command: Command, label: String, args: Array<out String>): Boolean {
+                    return try {
+                        val input = CommandInput(
+                                sender = sender,
+                                args = args ?: arrayOf(),
+                                isConsole = sender !is Player
+                        )
+                        when (wrappedCommand.execute(input)) {
+                            CommandResult.EXECUTED -> true
+                            CommandResult.INVALID_INPUT -> false
+                        }
+                    }
+                    catch (error: Exception) {
+                        sender.sendMessage("An internal error occurred performing your command")
+                        error.localizedMessage.let { message -> logger.fatal(message) }
+                        error.printStackTrace()
+                        true
                     }
                 }
-                catch (error: Exception) {
-                    sender.sendMessage("An internal error occurred performing your command")
-                    error.localizedMessage.let { message -> environment.log(LogLevel.FATAL, message) }
-                    true
-                }
             }
-            plugin.getCommand(alias)?.permission = command.permission
+            plugin.getCommand(alias).let {
+                it.setSuspendingExecutor(BridgedCommand(command))
+                it.permission = command.permission
+            }
         }
     }
 }
