@@ -1,29 +1,29 @@
 package com.projectcitybuild.platforms.spigot
 
-import com.projectcitybuild.core.contracts.EnvironmentProvider
-import com.projectcitybuild.core.contracts.Listenable
-import com.projectcitybuild.core.contracts.ListenerDelegatable
-import com.projectcitybuild.core.entities.LogLevel
+import com.github.shynixn.mccoroutine.registerSuspendingEvents
+import com.projectcitybuild.core.contracts.LoggerProvider
 import org.bukkit.event.HandlerList
+import org.bukkit.event.Listener
 import org.bukkit.plugin.java.JavaPlugin
-import java.lang.ref.WeakReference
 
 class SpigotListenerDelegate constructor(
-        private val plugin: WeakReference<JavaPlugin>,
-        private val environment: EnvironmentProvider
-    ): ListenerDelegatable {
+        private val plugin: JavaPlugin,
+        private val logger: LoggerProvider
+    ) {
 
-    override fun register(listener: Listenable<*>) {
-        val plugin = plugin.get() ?: throw Exception("Failed to register listener: Plugin is deallocated")
-        environment.log(LogLevel.VERBOSE, "Beginning listener registration...")
+    fun register(listener: Listener) {
+        logger.verbose("Beginning listener registration...")
 
-        plugin.server?.pluginManager?.registerEvents(listener, plugin).let {
-            environment.log(LogLevel.VERBOSE, "Registered listener: ${listener::class}")
-        }
+        plugin.server.pluginManager.registerSuspendingEvents(listener, plugin)
+        logger.verbose("Registered listener: ${listener::class.simpleName}")
+
+//        plugin.server?.pluginManager?.registerEvents(listener, plugin).let {
+//            logger.verbose("Registered listener: ${listener::class}")
+//        }
     }
 
-    override fun unregisterAll() {
+    fun unregisterAll() {
         HandlerList.unregisterAll()
-        environment.log(LogLevel.VERBOSE, "Unregistered all listeners")
+        logger.verbose("Unregistered all listeners")
     }
 }
