@@ -1,57 +1,69 @@
-package com.projectcitybuild
+package com.projectcitybuild.platforms.spigot
 
-import com.projectcitybuild.core.network.NetworkClients
+import com.projectcitybuild.core.network.APIRequestFactory
 import com.projectcitybuild.core.network.mojang.client.MojangClient
 import com.projectcitybuild.core.network.pcb.client.PCBClient
-import com.projectcitybuild.core.contracts.*
 import com.projectcitybuild.core.entities.PluginConfig
-import com.projectcitybuild.platforms.spigot.SpigotCommandDelegate
-import com.projectcitybuild.platforms.spigot.SpigotListenerDelegate
-import com.projectcitybuild.platforms.spigot.environment.SpigotEnvironment
-import com.projectcitybuild.platforms.spigot.environment.SpigotPlayerStore
-import com.projectcitybuild.platforms.spigot.environment.SpigotPluginHook
-import com.projectcitybuild.platforms.spigot.extensions.addDefault
+import com.projectcitybuild.core.network.APIClient
+import com.projectcitybuild.core.utilities.PlayerStore
+import com.projectcitybuild.modules.bans.CheckBanStatusAction
+import com.projectcitybuild.modules.ranks.SyncPlayerGroupAction
 import com.projectcitybuild.platforms.spigot.commands.*
+import com.projectcitybuild.platforms.spigot.environment.*
 import com.projectcitybuild.platforms.spigot.listeners.*
+import com.projectcitybuild.platforms.spigot.extensions.addDefault
 import org.bukkit.plugin.java.JavaPlugin
-import java.lang.ref.WeakReference
 
-class SpigotPlatform(plugin: JavaPlugin): PlatformBridgable {
+class SpigotPlatform: JavaPlugin() {
 
-    private var commandDelegate: CommandDelegatable? = null
-    private var listenerDelegate: ListenerDelegatable? = null
+    private val spigotLogger = SpigotLogger(logger = this.logger)
+    private val spigotConfig = SpigotConfig(config = this.config)
+    private val scheduler = SpigotScheduler(plugin = this)
+    private val apiClient = APIClient(plugin = this, logger = spigotLogger)
+    private val playerStore = PlayerStore()
+    private var playerStoreWrapper: SpigotPlayerStore? = null
+    private var permissionsManager: PermissionsManager? = null
+    private var commandDelegate: SpigotCommandDelegate? = null
+    private var listenerDelegate: SpigotListenerDelegate? = null
 
-    private var _networkClients: NetworkClients? = null
-    private val networkClients: NetworkClients
+    private var _apiRequestFactory: APIRequestFactory? = null
+    private val apiRequestFactory: APIRequestFactory
         get() {
-            if (_networkClients == null) {
-                _networkClients = createAPIProvider()
+            if (_apiRequestFactory == null) {
+                _apiRequestFactory = createAPIProvider()
             }
-            return _networkClients!!
+            return _apiRequestFactory!!
         }
 
-    private val weakRef = WeakReference(plugin)
-
-    override val environment: EnvironmentProvider by lazy {
-        SpigotEnvironment(
-            pluginRef = weakRef,
-            logger = plugin.logger,
-            playerStore = SpigotPlayerStore(plugin = weakRef).store,
-            config = plugin.config,
-            hooks = SpigotPluginHook()
+    private val syncPlayerGroupAction: SyncPlayerGroupAction by lazy {
+        SyncPlayerGroupAction(
+                permissionsManager!!,
+                apiRequestFactory,
+                apiClient,
+                spigotConfig,
+                spigotLogger
         )
+    }
+
+    private val checkBanStatusAction: CheckBanStatusAction by lazy {
+        CheckBanStatusAction(apiRequestFactory, apiClient)
     }
 
     override fun onEnable() {
         createDefaultConfig()
 
-        val commandDelegate = SpigotCommandDelegate(plugin = weakRef, environment = environment)
+        playerStoreWrapper = SpigotPlayerStore(plugin = this, store = playerStore)
+        permissionsManager = PermissionsManager()
+
+        val commandDelegate = SpigotCommandDelegate(plugin = this, logger = spigotLogger)
         registerCommands(delegate = commandDelegate)
         this.commandDelegate = commandDelegate
 
-        val listenerDelegate = SpigotListenerDelegate(plugin = weakRef, environment = environment)
+        val listenerDelegate = SpigotListenerDelegate(plugin = this, logger = spigotLogger)
         registerListeners(delegate = listenerDelegate)
         this.listenerDelegate = listenerDelegate
+
+        logger.info("PCBridge ready")
     }
 
     override fun onDisable() {
@@ -59,56 +71,81 @@ class SpigotPlatform(plugin: JavaPlugin): PlatformBridgable {
 
         commandDelegate = null
         listenerDelegate = null
+        permissionsManager = null
+
+        logger.info("PCBridge disabled")
     }
 
     private fun registerCommands(delegate: SpigotCommandDelegate) {
         arrayOf(
-                BanCommand(environment, networkClients),
-                UnbanCommand(environment, networkClients),
-                CheckBanCommand(environment, networkClients),
-                MuteCommand(environment),
-                UnmuteCommand(environment),
-                MaintenanceCommand(environment),
-                SyncCommand(environment, networkClients)
+                BanCommand(apiRequestFactory, apiClient),
+                UnbanCommand( apiRequestFactory, apiClient),
+                CheckBanCommand(scheduler, apiRequestFactory, apiClient, checkBanStatusAction),
+                MuteCommand(playerStore),
+                UnmuteCommand(playerStore),
+                MaintenanceCommand(),
+                SyncCommand(apiRequestFactory, apiClient, syncPlayerGroupAction),
+                SyncOtherCommand(syncPlayerGroupAction),
+                BoxCommand(apiRequestFactory, apiClient, spigotLogger)
         )
         .forEach { command -> delegate.register(command) }
     }
 
     private fun registerListeners(delegate: SpigotListenerDelegate) {
         arrayOf(
-                BanConnectionListener(environment, networkClients),
-                ChatListener(environment),
-                MaintenanceConnectListener(environment),
-                SyncRankLoginListener(environment, networkClients)
+                BanConnectionListener(apiRequestFactory, apiClient),
+                ChatListener(spigotConfig, playerStore, permissionsManager!!, spigotLogger),
+                MaintenanceConnectListener(spigotConfig),
+                SyncRankLoginListener(syncPlayerGroupAction),
+                AvailableBoxListener(apiRequestFactory, apiClient)
         )
         .forEach { listener -> delegate.register(listener) }
     }
 
     private fun createDefaultConfig() {
-        val plugin = weakRef.get() ?: throw Exception("Plugin reference lost")
+        config.addDefault(PluginConfig.SETTINGS.MAINTENANCE_MODE)
+        config.addDefault(PluginConfig.API.KEY)
+        config.addDefault(PluginConfig.API.BASE_URL)
+        config.addDefault(PluginConfig.GROUPS.GUEST)
+        config.addDefault(PluginConfig.GROUPS.TRUST_PRIORITY)
+        config.addDefault(PluginConfig.GROUPS.BUILD_PRIORITY)
+        config.addDefault(PluginConfig.GROUPS.DONOR_PRIORITY)
 
-        plugin.config.addDefault<PluginConfig.Settings.MAINTENANCE_MODE>()
-        plugin.config.addDefault<PluginConfig.API.KEY>()
-        plugin.config.addDefault<PluginConfig.API.BASE_URL>()
+        config.addDefault("groups.appearance.admin.display_name", "§4[Staff]")
+        config.addDefault("groups.appearance.admin.hover_name", "Administrator")
+        config.addDefault("groups.appearance.sop.display_name", "§c[Staff]")
+        config.addDefault("groups.appearance.sop.hover_name", "Senior Operator")
+        config.addDefault("groups.appearance.op.display_name", "§6[Staff]")
+        config.addDefault("groups.appearance.op.hover_name", "Operator")
+        config.addDefault("groups.appearance.moderator.display_name", "§e[Staff]")
+        config.addDefault("groups.appearance.moderator.hover_name", "Moderator")
 
-        plugin.config.options().copyDefaults(true)
-        plugin.saveConfig()
+        config.addDefault("groups.appearance.trusted+.hover_name", "Trusted+")
+        config.addDefault("groups.appearance.trusted.hover_name", "Trusted")
+        config.addDefault("groups.appearance.donator.hover_name", "Donor")
+        config.addDefault("groups.appearance.architect.hover_name", "Architect")
+        config.addDefault("groups.appearance.engineer.hover_name", "Engineer")
+        config.addDefault("groups.appearance.planner.hover_name", "Planner")
+        config.addDefault("groups.appearance.builder.hover_name", "Builder")
+        config.addDefault("groups.appearance.intern.hover_name", "Intern")
+
+        config.options().copyDefaults(true)
+        saveConfig()
     }
 
-    private fun createAPIProvider(): NetworkClients {
-        val isLoggingEnabled = environment.get(PluginConfig.API.IS_LOGGING_ENABLED()) as? Boolean
-                ?: throw Exception("Could not cast is_logging_enabled to Boolean")
+    private fun createAPIProvider(): APIRequestFactory {
+        val isLoggingEnabled = spigotConfig.get(PluginConfig.API.IS_LOGGING_ENABLED)
 
         val pcbClient = PCBClient(
-                authToken = environment.get(PluginConfig.API.KEY()) as? String
+                authToken = spigotConfig.get(PluginConfig.API.KEY) as? String
                         ?: throw Exception("Could not cast auth token to String"),
-                baseUrl = environment.get(PluginConfig.API.BASE_URL()) as? String
+                baseUrl = spigotConfig.get(PluginConfig.API.BASE_URL) as? String
                         ?: throw Exception("Could not cast base url to String"),
                 withLogging = isLoggingEnabled
         )
         val mojangClient = MojangClient(
                 withLogging = isLoggingEnabled
         )
-        return NetworkClients(pcb = pcbClient, mojang = mojangClient)
+        return APIRequestFactory(pcb = pcbClient, mojang = mojangClient)
     }
 }
