@@ -1,19 +1,19 @@
 package com.projectcitybuild.platforms.spigot.commands
 
-import com.projectcitybuild.core.contracts.CommandResult
+import com.projectcitybuild.core.entities.CommandResult
 import com.projectcitybuild.core.contracts.Commandable
-import com.projectcitybuild.core.contracts.EnvironmentProvider
 import com.projectcitybuild.core.entities.CommandInput
+import com.projectcitybuild.core.utilities.PlayerStore
 import com.projectcitybuild.platforms.spigot.extensions.getOnlinePlayer
 
 class MuteCommand(
-        private val environment: EnvironmentProvider
+        private val playerStore: PlayerStore
 ): Commandable {
 
     override val label = "mute"
     override val permission = "pcbridge.chat.mute"
 
-    override fun execute(input: CommandInput): CommandResult {
+    override suspend fun execute(input: CommandInput): CommandResult {
         if (!input.hasArguments) return CommandResult.INVALID_INPUT
 
         val targetPlayerName = input.args.first()
@@ -23,7 +23,7 @@ class MuteCommand(
             return CommandResult.INVALID_INPUT
         }
 
-        val player = environment.get(targetPlayer.uniqueId)
+        val player = playerStore.get(targetPlayer.uniqueId)
             ?: throw Exception("Player $targetPlayerName missing from cache")
 
         if (player.isMuted) {
@@ -32,7 +32,7 @@ class MuteCommand(
         }
 
         player.isMuted = true
-        environment.set(player)
+        playerStore.put(player.uuid, player)
 
         input.sender.sendMessage("${targetPlayer.name} has been muted")
         targetPlayer.sendMessage("You have been muted by ${input.sender.name}")
