@@ -1,7 +1,6 @@
 package com.projectcitybuild.platforms.spigot.environment
 
 import com.projectcitybuild.platforms.spigot.extensions.makeModel
-import com.projectcitybuild.core.contracts.PlayerStoreHoldable
 import com.projectcitybuild.core.utilities.PlayerStore
 import com.projectcitybuild.core.entities.Player
 import org.bukkit.configuration.file.YamlConfiguration
@@ -11,18 +10,12 @@ import org.bukkit.event.player.PlayerJoinEvent
 import org.bukkit.event.player.PlayerQuitEvent
 import org.bukkit.plugin.java.JavaPlugin
 import java.io.File
-import java.lang.ref.WeakReference
 import java.util.*
 
-class SpigotPlayerStore(val plugin: WeakReference<JavaPlugin>) : PlayerStoreHoldable, Listener, PlayerStore.PlayerStoreDelegate {
-    override val store: PlayerStore = PlayerStore()
-
+class SpigotPlayerStore(val plugin: JavaPlugin, val store: PlayerStore): Listener, PlayerStore.PlayerStoreDelegate {
     init {
-        val plugin = plugin.get()
-        if (plugin != null) {
-            plugin.server?.pluginManager?.registerEvents(this, plugin)
-            rebuildStore()
-        }
+        plugin.server?.pluginManager?.registerEvents(this, plugin)
+        rebuildStore()
         store.delegate = this
     }
 
@@ -44,14 +37,14 @@ class SpigotPlayerStore(val plugin: WeakReference<JavaPlugin>) : PlayerStoreHold
     private fun rebuildStore() {
         store.clear()
 
-        plugin.get()?.server?.onlinePlayers?.forEach { onlinePlayer ->
+        plugin.server?.onlinePlayers?.forEach { onlinePlayer ->
             val player = deserializeFromFile(uuid = onlinePlayer.uniqueId)
             store.put(onlinePlayer.uniqueId, player)
         }
     }
 
     private fun deserializeFromFile(uuid: UUID) : Player {
-        val folder = plugin.get()?.dataFolder?.absolutePath + File.separator + "players"
+        val folder = plugin.dataFolder?.absolutePath + File.separator + "players"
         val file = File(folder, "$uuid.yml")
 
         val reader = YamlConfiguration.loadConfiguration(file)
@@ -63,7 +56,7 @@ class SpigotPlayerStore(val plugin: WeakReference<JavaPlugin>) : PlayerStoreHold
     }
 
     private fun serializeToFile(uuid: UUID, player: Player? = null) {
-        val folder = plugin.get()?.dataFolder?.absolutePath + File.separator + "players"
+        val folder = plugin.dataFolder?.absolutePath + File.separator + "players"
         val file = File(folder, "$uuid.yml")
 
         val reader = YamlConfiguration.loadConfiguration(file)
