@@ -1,33 +1,30 @@
 package com.projectcitybuild.platforms.spigot.extensions
 
-import com.projectcitybuild.core.contracts.EnvironmentProvider
+import com.projectcitybuild.core.entities.Failure
+import com.projectcitybuild.core.entities.Success
 import com.projectcitybuild.core.extensions.toDashFormattedUUID
+import com.projectcitybuild.core.network.APIClient
 import com.projectcitybuild.modules.players.GetMojangPlayerAction
-import com.projectcitybuild.core.network.NetworkClients
+import com.projectcitybuild.core.network.APIRequestFactory
 import org.bukkit.Server
 import org.bukkit.entity.Player
 import java.util.*
 
-fun Server.getOfflinePlayer(
-        name: String,
-        environment: EnvironmentProvider,
-        networkClients: NetworkClients
+// TODO: move this into new Action
+suspend fun Server.getOfflinePlayer(
+    name: String,
+    apiRequestFactory: APIRequestFactory,
+    apiClient: APIClient
 ) : UUID? {
     val player = this.getOnlinePlayer(name)
     if (player != null) {
         return player.uniqueId
     }
-    val mojangPlayerAction = GetMojangPlayerAction(environment, networkClients)
-    val result = mojangPlayerAction.execute(playerName = name)
-    if (result is GetMojangPlayerAction.Result.SUCCESS) {
-        return UUID.fromString(result.player.uuid.toDashFormattedUUID())
+    val result = GetMojangPlayerAction(apiRequestFactory, apiClient).execute(playerName = name)
+    return when (result) {
+        is Success -> UUID.fromString(result.value.uuid.toDashFormattedUUID())
+        else -> return null
     }
-    if (result is GetMojangPlayerAction.Result.FAILED) {
-        when (result.reason) {
-            GetMojangPlayerAction.Failure.DESERIALIZE_FAILED -> throw Exception("Bad response from Mojang server when fetching UUID for offline player")
-        }
-    }
-    return null
 }
 
 fun Server.getOnlinePlayer(name: String) : Player? {
