@@ -1,208 +1,144 @@
 package com.projectcitybuild.platforms.spigot.listeners
 
-import com.projectcitybuild.core.contracts.EnvironmentProvider
-import com.projectcitybuild.core.contracts.Listenable
-import com.projectcitybuild.core.entities.BuildGroup
-import com.projectcitybuild.core.entities.DonorGroup
-import com.projectcitybuild.core.entities.LogLevel
-import com.projectcitybuild.core.entities.TrustGroup
-import net.luckperms.api.node.NodeType
+import com.projectcitybuild.core.contracts.ConfigProvider
+import com.projectcitybuild.core.contracts.LoggerProvider
+import com.projectcitybuild.core.entities.PluginConfig
+import com.projectcitybuild.core.utilities.PlayerStore
+import com.projectcitybuild.platforms.spigot.environment.PermissionsGroup
+import com.projectcitybuild.platforms.spigot.environment.PermissionsManager
 import net.md_5.bungee.api.ChatColor
 import net.md_5.bungee.api.chat.ComponentBuilder
 import net.md_5.bungee.api.chat.HoverEvent
 import net.md_5.bungee.api.chat.TextComponent
 import org.bukkit.event.EventHandler
 import org.bukkit.event.EventPriority
+import org.bukkit.event.Listener
 import org.bukkit.event.player.AsyncPlayerChatEvent
-import java.util.regex.Pattern
-import java.util.stream.Collectors
 
 
 /**
  * FIXME: Awful hacky, hardcoded stuff in here to save time
  */
 class ChatListener(
-        private val environment: EnvironmentProvider
-): Listenable<AsyncPlayerChatEvent> {
+        private val config: ConfigProvider,
+        private val playerStore: PlayerStore,
+        private val permissionsManager: PermissionsManager,
+        private val logger: LoggerProvider
+): Listener {
+
+    private val trustedGroups = config.get(PluginConfig.GROUPS.TRUST_PRIORITY)
+    private val builderGroups = config.get(PluginConfig.GROUPS.BUILD_PRIORITY)
+    private val donorGroups = config.get(PluginConfig.GROUPS.DONOR_PRIORITY)
+
+    init {
+        if (trustedGroups.isEmpty()) {
+            throw Exception("Trusted group config is empty. Did you forget to set this?")
+        }
+        if (builderGroups.isEmpty()) {
+            throw Exception("Builder group config is empty. Did you forget to set this?")
+        }
+        if (donorGroups.isEmpty()) {
+            throw Exception("Donor group config is empty. Did you forget to set this?")
+        }
+    }
 
     data class Group<GroupType>(
             val group: GroupType,
             val displayName: String,
             val hoverName: String
     )
-    private val trustGroupPriority: HashMap<TrustGroup, Int> = hashMapOf(
-            TrustGroup.GUEST to 0,
-            TrustGroup.MEMBER to 1,
-            TrustGroup.RETIRED to 2,
-            TrustGroup.TRUSTED to 3,
-            TrustGroup.TRUSTED_PLUS to 4,
-            TrustGroup.MODERATOR to 5,
-            TrustGroup.OPERATOR to 6,
-            TrustGroup.SENIOR_OPERATOR to 7,
-            TrustGroup.ADMINISTRATOR to 8
-    )
-    private val buildGroupPriority: HashMap<BuildGroup, Int> = hashMapOf(
-            BuildGroup.NONE to 0,
-            BuildGroup.INTERN to 1,
-            BuildGroup.BUILDER to 2,
-            BuildGroup.PLANNER to 3,
-            BuildGroup.ENGINEER to 4,
-            BuildGroup.ARCHITECT to 5
-    )
-
-    private fun highestTrustGroup(lhs: Group<TrustGroup>, rhs: Group<TrustGroup>): Group<TrustGroup> {
-        val leftPriority = trustGroupPriority[lhs.group] ?: 0
-        val rightPriority = trustGroupPriority[rhs.group] ?: 0
-
-        if (leftPriority > rightPriority) {
-            return lhs
-        }
-        return rhs
-    }
-
-    private fun highestBuildGroup(lhs: Group<BuildGroup>, rhs: Group<BuildGroup>): Group<BuildGroup> {
-        val leftPriority = buildGroupPriority[lhs.group] ?: -1
-        val rightPriority = buildGroupPriority[rhs.group] ?: -1
-
-        if (leftPriority > rightPriority) {
-            return lhs
-        }
-        return rhs
-    }
-
-    private val urlPattern = Regex("((?:(?:https?)://)?[\\w-_\\.]{2,})\\.([a-zA-Z]{2,3}(?:/\\S+)?)")
 
     @EventHandler(priority = EventPriority.HIGHEST)
-    override fun observe(event: AsyncPlayerChatEvent) {
-        val permissions = environment.permissions ?: throw Exception("Permission plugin is null")
-
+    fun onAsyncPlayerChatEvent(event: AsyncPlayerChatEvent) {
         // Mute player if necessary
-        val sendingPlayer = environment.get(event.player.uniqueId)
+        val sendingPlayer = playerStore.get(event.player.uniqueId)
         if (sendingPlayer?.isMuted == true) {
             event.isCancelled = true
             event.player.sendMessage("You cannot chat while muted")
+            logger.info("${event.player.displayName} tried to talk while muted")
             return
         }
 
         // Format user display name
-        val lpUser = permissions.userManager.getUser(event.player.uniqueId)
-                ?: throw Exception("Could not load user from LuckPerms")
+        val lpUser = permissionsManager.getUser(event.player.uniqueId)
+                ?: throw Exception("Could not load user from Permission plugin")
 
-        val groupNodes = lpUser.nodes.stream()
-                .filter(NodeType.INHERITANCE::matches)
-                .map(NodeType.INHERITANCE::cast)
-                .collect(Collectors.toSet())
+        val groupNodes = lpUser.groups()
 
-        val prefixes = lpUser.nodes.stream()
-                .filter(NodeType.PREFIX::matches)
-                .map(NodeType.PREFIX::cast)
-                .map { node -> node.metaValue }
-                .collect(Collectors.toSet())
-                .joinToString(separator = "")
+        var highestTrust: Pair<Int, PermissionsGroup>? = null
+        var highestBuild: Pair<Int, PermissionsGroup>? = null
+        var highestDonor: Pair<Int, PermissionsGroup>? = null
 
-        val suffixes = lpUser.nodes.stream()
-                .filter(NodeType.SUFFIX::matches)
-                .map(NodeType.SUFFIX::cast)
-                .map { node -> node.metaValue }
-                .collect(Collectors.toSet())
-                .joinToString(separator = "")
+        for (groupNode in groupNodes) {
+            val groupName = groupNode.name.toLowerCase()
 
-        // FIXME: cache so that this isn't performed every time a message is sent
-        var donorGroup = Group(DonorGroup.NONE, "", "")
-        var buildGroup = Group(BuildGroup.NONE, "", "")
-        var trustGroup = Group(TrustGroup.GUEST, "", "")
+            val trustIndex = trustedGroups.indexOf(groupName)
+            if (trustIndex != -1) {
+                if (highestTrust == null || trustIndex < highestTrust.first) {
+                    highestTrust = Pair(trustIndex, groupNode)
+                }
+            }
 
-        groupNodes.forEach { groupNode ->
-            val group = permissions.groupManager.getGroup(groupNode.groupName)
-            val displayName = group?.displayName ?: groupNode.groupName
+            val builderIndex = builderGroups.indexOf(groupName)
+            if (builderIndex != -1) {
+                if (highestBuild == null || builderIndex < highestBuild.first) {
+                    highestBuild = Pair(builderIndex, groupNode)
+                }
+            }
 
-            // FIXME: hardcoded for the sake of time, but this should all be from an API
-            val groupName = groupNode.groupName.toLowerCase()
-
-            when (groupName) {
-                "donator" -> donorGroup = Group(DonorGroup.DONOR, displayName, "Donor")
-                "member" -> {
-                    val newGroup = Group(TrustGroup.MEMBER, displayName, "")
-                    trustGroup = highestTrustGroup(trustGroup, newGroup)
-                }
-                "trusted" -> {
-                    val newGroup = Group(TrustGroup.TRUSTED, displayName, "Trusted")
-                    trustGroup = highestTrustGroup(trustGroup, newGroup)
-                }
-                "trusted+" -> {
-                    val newGroup = Group(TrustGroup.TRUSTED_PLUS, displayName, "Trusted+")
-                    trustGroup = highestTrustGroup(trustGroup, newGroup)
-                }
-                "moderator" -> {
-                    val newGroup = Group(TrustGroup.MODERATOR, "§e[Staff]", "Moderator")
-                    trustGroup = highestTrustGroup(trustGroup, newGroup)
-                }
-                "op" -> {
-                    val newGroup = Group(TrustGroup.OPERATOR, "§6[Staff]", "Operator")
-                    trustGroup = highestTrustGroup(trustGroup, newGroup)
-                }
-                "sop" -> {
-                    val newGroup = Group(TrustGroup.SENIOR_OPERATOR, "§c[Staff]", "Senior Operator")
-                    trustGroup = highestTrustGroup(trustGroup, newGroup)
-                }
-                "admin" -> {
-                    val newGroup = Group(TrustGroup.ADMINISTRATOR, "§4[Staff]", "Administrator")
-                    trustGroup = highestTrustGroup(trustGroup, newGroup)
-                }
-                "retired" -> {
-                    val newGroup = Group(TrustGroup.RETIRED, displayName, "Retired Staff")
-                    trustGroup = highestTrustGroup(trustGroup, newGroup)
-                }
-                "intern" -> {
-                    val newGroup = Group(BuildGroup.INTERN, displayName, "Intern")
-                    buildGroup = highestBuildGroup(buildGroup, newGroup)
-                }
-                "builder" -> {
-                    val newGroup = Group(BuildGroup.BUILDER, displayName, "Builder")
-                    buildGroup = highestBuildGroup(buildGroup, newGroup)
-                }
-                "planner" -> {
-                    val newGroup = Group(BuildGroup.PLANNER, displayName, "Planner")
-                    buildGroup = highestBuildGroup(buildGroup, newGroup)
-                }
-                "engineer" -> {
-                    val newGroup = Group(BuildGroup.ENGINEER, displayName, "Engineer")
-                    buildGroup = highestBuildGroup(buildGroup, newGroup)
-                }
-                "architect" -> {
-                    val newGroup = Group(BuildGroup.ARCHITECT, displayName, "Architect")
-                    buildGroup = highestBuildGroup(buildGroup, newGroup)
+            val donorIndex = donorGroups.indexOf(groupName)
+            if (donorIndex != -1) {
+                if (highestDonor == null || donorIndex < highestDonor.first) {
+                    highestDonor = Pair(donorIndex, groupNode)
                 }
             }
         }
 
         val groupTC = TextComponent()
-        if (donorGroup.group != DonorGroup.NONE) {
+        if (highestDonor != null) {
+            var hoverName = config.get(path = "groups.appearance.${highestDonor.second.name}.hover_name") as? String
+            var displayName = config.get(path = "groups.appearance.${highestDonor.second.name}.display_name") as? String
+            if (displayName.isNullOrBlank()) {
+                displayName = highestDonor.second.getDisplayName(permissionsManager)
+            }
             TextComponent
-                    .fromLegacyText(donorGroup.displayName)
+                    .fromLegacyText(displayName)
                     .forEach { c ->
-                        val hoverEvent = HoverEvent(HoverEvent.Action.SHOW_TEXT, ComponentBuilder(donorGroup.hoverName).create())
+                        if (hoverName.isNullOrBlank()) return
+                        val hoverEvent = HoverEvent(HoverEvent.Action.SHOW_TEXT, ComponentBuilder(hoverName).create())
                         c.hoverEvent = hoverEvent
                         groupTC.addExtra(c)
                     }
         }
-        if (trustGroup.group != TrustGroup.GUEST && trustGroup.displayName.isNotBlank()) {
+        if (highestTrust != null) {
+            var hoverName = config.get(path = "groups.appearance.${highestTrust.second.name}.hover_name") as? String
+            var displayName = config.get(path = "groups.appearance.${highestTrust.second.name}.display_name") as? String
+            if (displayName.isNullOrBlank()) {
+                displayName = highestTrust.second.getDisplayName(permissionsManager)
+            }
             TextComponent
-                    .fromLegacyText(trustGroup.displayName)
-                    .forEach { c ->
-                        val hoverEvent = HoverEvent(HoverEvent.Action.SHOW_TEXT, ComponentBuilder(trustGroup.hoverName).create())
-                        c.hoverEvent = hoverEvent
-                        groupTC.addExtra(c)
-                    }
+                .fromLegacyText(displayName)
+                .forEach { c ->
+                    if (hoverName.isNullOrBlank()) return
+                    val hoverEvent = HoverEvent(HoverEvent.Action.SHOW_TEXT, ComponentBuilder(hoverName).create())
+                    c.hoverEvent = hoverEvent
+                    groupTC.addExtra(c)
+                }
         }
-        if (buildGroup.group != BuildGroup.NONE && buildGroup.displayName.isNotBlank()) {
+        if (highestBuild != null) {
+            var hoverName = config.get(path = "groups.appearance.${highestBuild.second.name}.hover_name") as? String
+            var displayName = config.get(path = "groups.appearance.${highestBuild.second.name}.display_name") as? String
+            if (displayName.isNullOrBlank()) {
+                displayName = highestBuild.second.getDisplayName(permissionsManager)
+            }
             TextComponent
-                    .fromLegacyText(buildGroup.displayName)
-                    .forEach { c ->
-                        val hoverEvent = HoverEvent(HoverEvent.Action.SHOW_TEXT, ComponentBuilder(buildGroup.hoverName).create())
-                        c.hoverEvent = hoverEvent
-                        groupTC.addExtra(c)
-                    }
+                .fromLegacyText(displayName)
+                .forEach { c ->
+                    if (hoverName.isNullOrBlank()) return
+                    val hoverEvent = HoverEvent(HoverEvent.Action.SHOW_TEXT, ComponentBuilder(hoverName).create())
+                    c.hoverEvent = hoverEvent
+                    groupTC.addExtra(c)
+                }
         }
 
         val messageTC = TextComponent.fromLegacyText(event.message)
@@ -214,8 +150,8 @@ class ChatListener(
         colonTC.color = ChatColor.RESET
 
         // Dynamic text from other plugins (eg. LuckyPerms) contains legacy color codes
-        val prefixTC = TextComponent.fromLegacyText(prefixes)
-        val suffixTC = TextComponent.fromLegacyText(suffixes)
+        val prefixTC = TextComponent.fromLegacyText(lpUser.prefixes())
+        val suffixTC = TextComponent.fromLegacyText(lpUser.suffixes())
         val displayNameTC = TextComponent.fromLegacyText(event.player.displayName)
 
         val textComponent = TextComponent()
@@ -234,15 +170,7 @@ class ChatListener(
         }
 
         // Messages sent to users don't appear in console, so we have to log it manually
-        environment.log(LogLevel.INFO, "<${event.player.displayName}> ${event.message}")
-
-//        // Terrible hack to hide the message without cancelling it.
-//        //
-//        // Normally we would cancel the event, but DiscordSRV is competing for the
-//        // highest priority and won't see the message if we cancel it.
-//        //
-//        // This will probably degrade performance...
-//        event.recipients.clear()
+        logger.info("<${event.player.displayName}> ${event.message}")
     }
 
 }
